first version of castor3d output. to be tested....

+++ src/readstate/readstate_vars.f90
@@ -34,11 +34,11 @@ PUBLIC
   CLASS(t_base),ALLOCATABLE :: X1_base_r              !! container for base of X1
   CLASS(t_base),ALLOCATABLE :: X2_base_r              !! container for base of X2
   CLASS(t_base),ALLOCATABLE :: LA_base_r              !! container for base of LA
-  REAL,ALLOCATABLE          :: X1_r(:,:)              !! spline x fourier coefs of solution X1
-  REAL,ALLOCATABLE          :: X2_r(:,:)              !! spline x fourier coefs of solution X2
-  REAL,ALLOCATABLE          :: LA_r(:,:)              !! spline x fourier coefs of solution LA 
-  REAL,ALLOCATABLE          :: profiles_1d(:,:)       !! spline coefficients for 1d profiles (using X1_base...needs to be improved!)
-  REAL                      :: a_minor,r_major,volume !! scalars: average minor and major radius, total volume
+  REAL(wp),ALLOCATABLE      :: X1_r(:,:)              !! spline x fourier coefs of solution X1
+  REAL(wp),ALLOCATABLE      :: X2_r(:,:)              !! spline x fourier coefs of solution X2
+  REAL(wp),ALLOCATABLE      :: LA_r(:,:)              !! spline x fourier coefs of solution LA 
+  REAL(wp),ALLOCATABLE      :: profiles_1d(:,:)       !! spline coefficients for 1d profiles (using X1_base...needs to be improved!)
+  REAL(wp)                  :: a_minor,r_major,volume !! scalars: average minor and major radius, total volume
 !===================================================================================================================================
 END MODULE MODgvec_ReadState_Vars
